added full-screen option, fixed various feedback items

+++ url_screenshot/icon_url_screenshot/actions/screenshot/action.py
@@ -14,6 +14,9 @@ from selenium.webdriver.chrome.options import Options
 class Screenshot(komand.Action):
 
     def __init__(self):
+        self.CHROME_PATH = '/usr/bin/chromium'
+        self.CHROMEDRIVER_PATH = '/usr/bin/chromedriver'
+        self.WINDOW_SIZE = "1920,1080"
         super(self.__class__, self).__init__(
                 name='screenshot',
                 description=Component.DESCRIPTION,
@@ -21,32 +24,36 @@ class Screenshot(komand.Action):
                 output=ScreenshotOutput())
 
     def run(self, params={}):
-        CHROME_PATH = '/usr/bin/chromium'
-        CHROMEDRIVER_PATH = '/usr/bin/chromedriver'
-        WINDOW_SIZE = "1920,1080"
 
-        name = "/tmp/{}{}".format(str(uuid.uuid1()), ".png")
+        name = f"/tmp/{str(uuid.uuid1())}.png"
 
         chrome_options = Options()
         chrome_options.add_argument("--headless")
-        chrome_options.add_argument("--window-size=%s" % WINDOW_SIZE)
+        chrome_options.add_argument("--window-size=%s" % self.WINDOW_SIZE)
         chrome_options.add_argument("--no-sandbox")
-        chrome_options.binary_location = CHROME_PATH
+        chrome_options.binary_location = self.CHROME_PATH
 
         url = params.get(Input.URL, "")
         delay = params.get(Input.DELAY, 0)
         if not url.lower().startswith('http'):
             raise PluginException(cause='Input Error:', assistance='URLs need to start with "http"')
-        driver = webdriver.Chrome(executable_path=CHROMEDRIVER_PATH, options=chrome_options)
+        driver = webdriver.Chrome(executable_path=self.CHROMEDRIVER_PATH, options=chrome_options)
         driver.get(url)
         time.sleep(delay)
-        driver.save_screenshot(name)
-        driver.close()
+        full_page = params.get(Input.FULL_PAGE, False)
+        if full_page:
+            body = driver.find_element_by_tag_name('body')
+            element_png = body.screenshot_as_png
+            img_str = base64.b64encode(element_png).decode('ascii')
+            return {Output.SCREENSHOT: img_str}
+        else:
+            driver.save_screenshot(name)
+            driver.close()
 
-        with open(name, "rb") as file:
-            contents = file.read()
-            file_str = base64.b64encode(contents).decode('ascii')
+            with open(name, "rb") as file:
+                contents = file.read()
+                file_str = base64.b64encode(contents).decode('ascii')
 
-        os.remove(name)
+            os.remove(name)
 
-        return {Output.SCREENSHOT: file_str}
+            return {Output.SCREENSHOT: file_str}
